dashboard/requirements.txt
Flask==1.1.1
docker==4.1.0
requests==2.22.0
python-gitlab==1.13.0
python-jenkins==1.5.0